requirements.txt
Django>=4.0
djangorestframework>=3.13
djangorestframework-simplejwt>=5.0
psycopg2-binary>=2.9
django-celery-results>=2.3redis
celery==5.3.0
redis==5.0.0

